+++ frontend/src/components/CANs/CanCard/CanCard.test.jsx
@@ -7,7 +7,7 @@ import { getPortfolioCansFundingDetails } from "../../../api/getCanFundingSummar
 vi.mock("../../../api/getCanFundingSummary");
 
 // Mock the ResponsiveDonutWithInnerPercent component
-vi.mock("../../UI/ResponsiveDonutWithInnerPercent/ResponsiveDonutWithInnerPercent", () => ({
+vi.mock("../../UI/DataViz/ResponsiveDonutWithInnerPercent/ResponsiveDonutWithInnerPercent", () => ({
     ResponsiveDonutWithInnerPercent: () => <div data-testid="mock-donut-chart" />
 }));
