+++ frontend/src/components/Portfolios/PortfolioSummaryCards/PortfolioCarryForwardFunding.jsx
@@ -1,5 +1,5 @@
 import { useSelector } from "react-redux";
-import CurrencySummaryCard from "../../UI/CurrencySummaryCard/CurrencySummaryCard";
+import CurrencySummaryCard from "../../UI/SummaryCard/CurrencySummaryCard";
 
 const PortfolioCarryForwardFunding = () => {
     const portfolioBudget = useSelector((state) => state.portfolioBudgetSummary.portfolioBudget);
