+++ frontend/src/components/Portfolios/PortfolioSummaryCards/PortfolioFundingTotal.jsx
@@ -1,11 +1,11 @@
-import React from "react";
-import { useSelector } from "react-redux";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faCircle } from "@fortawesome/free-solid-svg-icons";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import React from "react";
 import CurrencyFormat from "react-currency-format";
-import CurrencySummaryCard from "../../UI/CurrencySummaryCard";
-import LineGraph from "../../UI/DataViz/LineGraph";
+import { useSelector } from "react-redux";
 import { calculatePercent } from "../../../helpers/utils";
+import LineGraph from "../../UI/DataViz/LineGraph";
+import CurrencySummaryCard from "../../UI/SummaryCard/CurrencySummaryCard";
 import Tag from "../../UI/Tag/Tag";
 
 const PortfolioFundingTotal = () => {
